# edubridge_backend/docs/api/dashboard/openapi-dashboard-v1.yaml
openapi: 3.0.0
info:
  title: EduBridge Dashboard API
  version: 1.0.0
  description: API specifications for the EduBridge school management dashboard panel.
servers:
  - url: http://alpha.edubridge.test/api/v1
    description: Local development environment
  - url: https://{school_subdomain}.api.edubridge.com/api/v1
    description: Production API URL
    variables:
      school_subdomain:
        default: alpha
        description: Subdomain assigned to the tenant school
security:
  - BearerAuth: []
paths:
...
  /auth/device-sessions/{deviceSession}/revoke:
    post:
      summary: Revoke a specific device session
      operationId: revokeDeviceSession
      tags:
        - Authentication
      parameters:
        - name: deviceSession
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Device session revoked
  /auth/device/push-token:
    put:
      summary: Update device FCM token
      operationId: updatePushToken
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, platform]
              properties:
                token:
                  type: string
                platform:
                  type: string
                  enum: [android, ios, web]
      responses:
        '204':
          description: FCM Token saved
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...